reorganized file structure by resource and by construct, added css reset and base
