+++ chris_lee/client/app/css/base.css
@@ -1,0 +1,16 @@
+body {
+	background-color: #2C5371;
+	font-family: 'Open Sans', sans-serif;
+	padding: 20px;
+}
+h1 {
+	font-size: 30px;
+	padding: 20px;
+}
+h2 {
+	font-size: 20px;
+	padding: 20px;
+}
+ol, ul {
+	padding: 20px;
+}
