+++ chris_lee/client/app/js/ct/index.js
@@ -1,0 +1,3 @@
+module.exports = function(app) {
+  require('./controllers/ct_controller')(app);
+};
