+++ chris_lee/client/app/js/services/index.js
@@ -1,0 +1,3 @@
+module.exports = function(app) {
+  require('./resource_service')(app);
+};
