+++ chris_lee/client/app/js/t/index.js
@@ -1,0 +1,3 @@
+module.exports = function(app) {
+  require('./controllers/t_controller')(app);
+};
